Refactoring core logic for fixing into idflakies-core and preserve maven specific stuff under idflakies-maven-plugin
