+++ idflakies-core/src/main/java/edu/illinois/cs/dt/tools/utility/BuildCommands.java
@@ -1,0 +1,6 @@
+package edu.illinois.cs.dt.tools.utility;
+
+public abstract class BuildCommands {
+
+    public abstract void install();
+}
